rest/serializers.py
# -*- coding: utf-8 -*-
from rest_framework import serializers

from blog.models import Post


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'
        read_only_fields = ('cdate', 'edate',)
