requirements.txt
django>=1.10,<1.11
djangorestframework>=3.5.0,<3.6
markdown
